requirements.txt
# This is a set of known working versions for inference, documented for a
# distant future.
# We recommend following the requirements section in our README.md instead.
# einops==0.8.0
# numpy==1.26.4
# rotary_embedding_torch==0.6.4
# soxr==0.3.7
# torch==2.3.1
# torchaudio==2.3.1
# tqdm==4.66.4

tqdm 
einops 
soxr 
rotary-embedding-torch 
pytorch_lightning
pandas
mir_eval
wandb
nvitop
./src/madmom
./src/librosa
-e .
-e ./src/temgo
-e ./src/tempnetic